requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
filelock==3.0.12
googletrans==2.4.0
idna==2.8
importlib-metadata==0.19
more-itertools==7.2.0
packaging==19.1
pdf2image==1.7.0
Pillow==6.1.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.2
pytesseract==0.3.0
requests==2.22.0
rope==0.14.0
six==1.12.0
toml==0.10.0
tox==3.13.2
urllib3==1.25.3
virtualenv==16.7.4
zipp==0.6.0
